validato message edit data

+++ routes/chat.js
@@ -1,6 +1,8 @@
 const express = require('express');
 const router = express.Router();
 const connection = require('../mysqlConnection');
+const { validationResult } = require('express-validator');
+const validationCheck = require('../public/javascripts/validation/chat-text-edit/validation');
 
 //各ルームのメッセージ詳細ページ
 router.get('/:room_id/text/:text_id/account/:user_id/show', function(req, res, next) {
@@ -63,23 +65,30 @@ router.get('/:room_id/text/:text_id/account/:user_id/edit', function(req, res, n
   });
 });
 
-router.post('/:room_id/text/:text_id/account/:user_id/edit', (req, res, next) => {
+router.post('/:room_id/text/:text_id/account/:user_id/edit', validationCheck, (req, res, next) => {
   const room_id = req.params.room_id;
   const text_id = req.params.text_id;
+  const user_id = req.params.user_id;
   const text = req.body.text;
   const num = req.body.edit;
-  if (num == 1) {
-    const query = 'DELETE FROM message WHERE message_id = ?';
-    connection.query(query, [text_id], (err, rows) => {
-      if (err) throw err;
-      res.redirect('/chat/' + room_id);
-    });
-  } 
-  if (num == 2) {
-    connection.query('UPDATE message SET text = ? WHERE message_id = ?', [text, text_id], (err, rows) => {
-      if (err) throw err;
-      res.redirect('/chat/room/' + req.params.room_id + '/text/' + req.params.text_id + '/account/' + req.params.user_id);
-    });
+  const validationError = validationResult(req);
+  if (!validationError.isEmpty()) {
+    console.info(validationError.errors);
+    res.redirect('/' + room_id + '/text/' + text_id + '/account/' + user_id + '/edit');
+  } else {
+    if (num == 1) {
+      const query = 'DELETE FROM message WHERE message_id = ?';
+      connection.query(query, [text_id], (err, rows) => {
+        if (err) throw err;
+        res.redirect('/chat/' + room_id);
+      });
+    } 
+    if (num == 2) {
+      connection.query('UPDATE message SET text = ? WHERE message_id = ?', [text, text_id], (err, rows) => {
+        if (err) throw err;
+        res.redirect('/chat/room/' + req.params.room_id + '/text/' + req.params.text_id + '/account/' + req.params.user_id);
+      });
+    }
   }
 });
